+++ Week02/ReactViteTS/src/App.css
@@ -7,4 +7,7 @@
 body {
   background-color: rgb(196, 233, 255);
   font-family: sans-serif;
+  display: flex;
+  justify-content: center;
+  align-items: center;
 }
