+++ Week02/ReactViteTS/src/main.tsx
@@ -2,12 +2,9 @@ import { StrictMode } from "react";
 import { createRoot } from "react-dom/client";
 import "./index.css";
 import App from "./App.tsx";
-import { CounterProvider } from "./context/CounterProvider.tsx";
 
 createRoot(document.getElementById("root")!).render(
   <StrictMode>
-    <CounterProvider>
-      <App />
-    </CounterProvider>
+    <App />
   </StrictMode>
 );
